cache GitOps workshop event details when server starts
